refactor : move from config to controller
