+++ src/main/java/com/ctrls/auto_enter_view/controller/JobPostingImageController.java
@@ -1,4 +1,4 @@
-package com.ctrls.auto_enter_view.config;
+package com.ctrls.auto_enter_view.controller;
 
 import com.ctrls.auto_enter_view.service.JobPostingImageService;
 import lombok.RequiredArgsConstructor;
